learn 7 - complete bank simulation

+++ backend/src/main/java/com/nurs/backend/dto/TransactionRequest.java
@@ -1,5 +1,7 @@
 package com.nurs.backend.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
@@ -12,7 +14,7 @@ public class TransactionRequest {
     @NotNull(message = "amount is required")
     @Min(1)
     @Max(100_000_000)
-    private Long amount;
+    private BigDecimal amount;
 
     @NotBlank(message = "destinationId is required")
     @Size(min = 10, max = 10)
