+++ backend/src/main/java/com/nurs/backend/dto/UserResponse.java
@@ -1,5 +1,7 @@
 package com.nurs.backend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Value;
 
 @Value
@@ -8,5 +10,5 @@ public class UserResponse {
   private String accountId;
   private String roles;
   private String username;
-  private Long balance;
+  private BigDecimal balance;
 }
